+++ packages/network-controller/src/NetworkController.ts
@@ -38,7 +38,7 @@ export type Block = {
   baseFeePerGas?: string;
 };
 
-export type NetworkProperties = {
+export type NetworkDetails = {
   isEIP1559Compatible?: boolean;
 };
 
@@ -54,7 +54,7 @@ export type NetworkState = {
   network: string;
   isCustomNetwork: boolean;
   providerConfig: ProviderConfig;
-  properties: NetworkProperties;
+  networkDetails: NetworkDetails;
 };
 
 const LOCALHOST_RPC_URL = 'http://localhost:8545';
@@ -110,7 +110,7 @@ export const defaultState: NetworkState = {
   network: 'loading',
   isCustomNetwork: false,
   providerConfig: { type: MAINNET, chainId: NetworksChainId.mainnet },
-  properties: { isEIP1559Compatible: false },
+  networkDetails: { isEIP1559Compatible: false },
 };
 
 /**
@@ -141,7 +141,7 @@ export class NetworkController extends BaseControllerV2<
           persist: true,
           anonymous: false,
         },
-        properties: {
+        networkDetails: {
           persist: true,
           anonymous: false,
         },
@@ -202,7 +202,7 @@ export class NetworkController extends BaseControllerV2<
   private refreshNetwork() {
     this.update((state) => {
       state.network = 'loading';
-      state.properties = {};
+      state.networkDetails = {};
     });
     const { rpcTarget, type, chainId, ticker } = this.state.providerConfig;
     this.initializeProvider(type, rpcTarget, chainId, ticker);
@@ -400,34 +400,40 @@ export class NetworkController extends BaseControllerV2<
     this.refreshNetwork();
   }
 
-  getEIP1559Compatibility() {
-    const { properties = {} } = this.state;
+  #getLatestBlock(): Promise<Block> {
+    return new Promise((resolve, reject) => {
+      this.ethQuery.sendAsync(
+        { method: 'eth_getBlockByNumber', params: ['latest', false] },
+        (error: Error, block: Block) => {
+          if (error) {
+            reject(error);
+          } else {
+            resolve(block);
+          }
+        },
+      );
+    });
+  }
 
-    if (!properties.isEIP1559Compatible) {
-      if (typeof this.ethQuery?.sendAsync !== 'function') {
-        return Promise.resolve(true);
-      }
-      return new Promise((resolve, reject) => {
-        this.ethQuery.sendAsync(
-          { method: 'eth_getBlockByNumber', params: ['latest', false] },
-          (error: Error, block: Block) => {
-            if (error) {
-              reject(error);
-            } else {
-              const isEIP1559Compatible =
-                typeof block.baseFeePerGas !== 'undefined';
-              if (properties.isEIP1559Compatible !== isEIP1559Compatible) {
-                this.update((state) => {
-                  state.properties.isEIP1559Compatible = isEIP1559Compatible;
-                });
-              }
-              resolve(isEIP1559Compatible);
-            }
-          },
-        );
+  async getEIP1559Compatibility() {
+    const { networkDetails = {} } = this.state;
+
+    if (
+      networkDetails.isEIP1559Compatible ||
+      typeof this.ethQuery?.sendAsync !== 'function'
+    ) {
+      return true;
+    }
+
+    const latestBlock = await this.#getLatestBlock();
+    const isEIP1559Compatible =
+      typeof latestBlock.baseFeePerGas !== 'undefined';
+    if (networkDetails.isEIP1559Compatible !== isEIP1559Compatible) {
+      this.update((state) => {
+        state.networkDetails.isEIP1559Compatible = isEIP1559Compatible;
       });
     }
-    return Promise.resolve(true);
+    return isEIP1559Compatible;
   }
 }